mnpbsnl/requirements.txt
# django>=3.2,<4.0
django==3.1.5
gunicorn
requests
django-dotenv
psycopg2-binary
django-storages
boto3
